+++ src/main/java/com/b1n_ry/yigd/Yigd.java
@@ -215,11 +215,12 @@ public class Yigd implements ModInitializer, DedicatedServerModInitializer {
         if (FabricLoader.getInstance().isModLoaded("inventorio")) {
             apiMods.add(new InventorioCompat());
         }
-        if (FabricLoader.getInstance().isModLoaded("travelersbackpack")) {
+        if (FabricLoader.getInstance().isModLoaded("travelersbackpack") && !TravelersBackpackCompat.isTrinketIntegrationOn()) {
             apiMods.add(new TravelersBackpackCompat());
         }
         if (FabricLoader.getInstance().isModLoaded("apoli")) {
             apiMods.add(new OriginsCompat());
+            miscCompatMods.add("apoli");
         }
         apiMods.addAll(FabricLoader.getInstance().getEntrypoints("yigd", YigdApi.class));