+++ src/main/java/com/b1n_ry/yigd/core/ClientPacketReceivers.java
@@ -3,14 +3,14 @@ package com.b1n_ry.yigd.core;
 import com.b1n_ry.yigd.client.gui.GraveSelectScreen;
 import com.b1n_ry.yigd.client.gui.GraveViewScreen;
 import com.b1n_ry.yigd.client.gui.PlayerSelectScreen;
+import com.mojang.authlib.GameProfile;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtHelper;
+import net.minecraft.util.math.BlockPos;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 
 public class ClientPacketReceivers {
     public static void register() {
@@ -23,70 +23,74 @@ public class ClientPacketReceivers {
             GraveViewScreen.Permissions.delete = buf.readBoolean();
             GraveViewScreen.Permissions.rob = buf.readBoolean();
 
+            boolean unlocked = buf.readBoolean();
+            boolean showGraveRobber = buf.readBoolean();
+
             DeadPlayerData data = DeadPlayerData.fromNbt(nbtData);
 
-            GraveViewScreen.unlockedGraves.clear();
-            int unlockedGraveSize = buf.readInt();
-            for (int i = 0; i < unlockedGraveSize; i++) {
-                UUID uuid = buf.readUuid();
-                GraveViewScreen.unlockedGraves.add(uuid);
-            }
-
             client.execute(() -> {
-                GraveViewScreen screen = new GraveViewScreen(data, null);
+                GraveViewScreen screen = new GraveViewScreen(data, unlocked, showGraveRobber, client.currentScreen);
                 MinecraftClient.getInstance().setScreen(screen);
             });
         });
         ClientPlayNetworking.registerGlobalReceiver(PacketIdentifiers.PLAYER_GRAVES_GUI, (client, handler, buf, responseSender) -> {
             if (client == null) return;
 
-            UUID userId = buf.readUuid();
-            DeadPlayerData data = DeadPlayerData.fromNbt(buf.readNbt());
-            boolean isUnlocked = buf.readBoolean();
+            NbtCompound profileNbt = buf.readNbt();
+            if (profileNbt == null) {
+                return;
+            }
+            GameProfile profile = NbtHelper.toGameProfile(profileNbt);
+            int graveCount = buf.readInt();
+            Map<UUID, GraveSelectScreen.GraveGuiInfo> graves = new HashMap<>();
+            for (int i = 0; i < graveCount; i++) {
+                UUID graveId = buf.readUuid();
+                BlockPos pos = buf.readBlockPos();
+                String dimensionName = buf.readString();
+                int itemCount = buf.readInt();
+                int levelCount = buf.readInt();
+                byte availability = buf.readByte();
+
+                GraveSelectScreen.GraveGuiInfo info = new GraveSelectScreen.GraveGuiInfo(pos, dimensionName, itemCount, levelCount, availability);
+
+                graves.put(graveId, info);
+            }
 
             client.execute(() -> {
-                if (client.currentScreen instanceof GraveSelectScreen openScreen) {
-                    openScreen.addData(userId, data);
-                } else {
-                    GraveViewScreen.unlockedGraves.clear();
-
-                    GraveSelectScreen screen = new GraveSelectScreen(List.of(data), 1, client.currentScreen);
-                    client.setScreen(screen);
-                }
-
-                if (isUnlocked) GraveViewScreen.unlockedGraves.add(data.id);
+                GraveSelectScreen screen = new GraveSelectScreen(profile, graves, 1, client.currentScreen);
+                client.setScreen(screen);
             });
         });
         ClientPlayNetworking.registerGlobalReceiver(PacketIdentifiers.ALL_PLAYER_GRAVES, (client, handler, buf, responseSender) -> {
             if (client == null) return;
 
-            UUID userId = buf.readUuid();
-            DeadPlayerData data = DeadPlayerData.fromNbt(buf.readNbt());
-            boolean isUnlocked = buf.readBoolean();
+            List<GameProfile> gameProfiles = new ArrayList<>();
+            Map<UUID, List<Byte>> graveAvailabilities = new HashMap<>();
+
+            int playerCount = buf.readInt();
+            for (int i = 0; i < playerCount; i++) {
+                NbtCompound profileNbt = buf.readNbt();
+                if (profileNbt != null) {
+                    GameProfile profile = NbtHelper.toGameProfile(profileNbt);
+
+                    if (profile != null) {
+                        List<Byte> bytes = new ArrayList<>();
+                        int graveCount = buf.readInt();
+                        for (int n = 0; n < graveCount; n++) {
+                            byte availability = buf.readByte();
+                            bytes.add(availability);
+                        }
+
+                        gameProfiles.add(profile);
+                        graveAvailabilities.put(profile.getId(), bytes);
+                    }
+                }
+            }
 
             client.execute(() -> {
-                if (client.currentScreen instanceof PlayerSelectScreen openScreen) {
-                    openScreen.addData(userId, data);
-                } else {
-                    GraveViewScreen.unlockedGraves.clear();
-
-                    Map<UUID, List<DeadPlayerData>> dataMap = new HashMap<>();
-                    dataMap.put(userId, List.of(data));
-
-                    PlayerSelectScreen screen = new PlayerSelectScreen(dataMap, 1);
-                    client.setScreen(screen);
-                }
-
-                if (isUnlocked) GraveViewScreen.unlockedGraves.add(data.id);
+                PlayerSelectScreen screen = new PlayerSelectScreen(gameProfiles, graveAvailabilities, 1);
+                client.setScreen(screen);
             });
-        });
-        ClientPlayNetworking.registerGlobalReceiver(PacketIdentifiers.GUI_CONFIGS, (client, handler, buf, responseSender) -> {
-            GraveViewScreen.Permissions.giveKey = buf.readBoolean();
-            GraveViewScreen.Permissions.toggleLock = buf.readBoolean();
-            GraveViewScreen.showGraveRobber = buf.readBoolean();
-            GraveViewScreen.Permissions.restore = buf.readBoolean();
-            GraveViewScreen.Permissions.delete = buf.readBoolean();
-            GraveViewScreen.Permissions.rob = buf.readBoolean();
         });
     }
 }