+++ src/main/java/com/b1n_ry/yigd/core/ServerPacketReceivers.java
@@ -1,11 +1,17 @@
 package com.b1n_ry.yigd.core;
 
 import com.b1n_ry.yigd.Yigd;
+import com.b1n_ry.yigd.api.YigdApi;
 import com.b1n_ry.yigd.config.PriorityInventoryConfig;
+import com.b1n_ry.yigd.config.YigdConfig;
 import com.b1n_ry.yigd.item.KeyItem;
 import com.mojang.authlib.GameProfile;
+import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
-import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtHelper;
+import net.minecraft.network.PacketByteBuf;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.TranslatableText;
 
@@ -79,6 +85,91 @@ public class ServerPacketReceivers {
                 DeathInfoManager.INSTANCE.unlockedGraves.add(graveId);
             }
             DeathInfoManager.INSTANCE.markDirty();
+        });
+
+        // Request listeners
+        ServerPlayNetworking.registerGlobalReceiver(PacketIdentifiers.PLAYER_GRAVES_GUI, (server, player, handler, serverBuf, responseSender) -> {
+            NbtCompound nbtProfile = serverBuf.readNbt();
+
+            if (nbtProfile == null) return;
+
+            GameProfile profile = NbtHelper.toGameProfile(nbtProfile);
+            if (profile == null) return;
+
+            List<DeadPlayerData> graveData = DeathInfoManager.INSTANCE.data.get(profile.getId());
+            if (graveData == null) return;
+
+            PacketByteBuf buf = PacketByteBufs.create();
+            buf.writeNbt(NbtHelper.writeGameProfile(new NbtCompound(), profile));
+            buf.writeInt(graveData.size());
+            for (DeadPlayerData grave : graveData) {
+                int itemCount = 0;
+                for (ItemStack stack : grave.inventory) {
+                    if (!stack.isEmpty()) itemCount++;
+                }
+                for (int i = 0; i < Yigd.apiMods.size(); i++) {
+                    YigdApi yigdApi = Yigd.apiMods.get(i);
+
+                    itemCount += yigdApi.getInventorySize(grave.modInventories.get(i));
+                }
+
+                int points = grave.xp;
+                int i;
+                for (i = 0; points >= 0; i++) {
+                    if (i < 16) points -= (2 * i) + 7;
+                    else if (i < 31) points -= (5 * i) - 38;
+                    else points -= (9 * i) - 158;
+                }
+
+                buf.writeUuid(grave.id);
+                buf.writeBlockPos(grave.gravePos);
+                buf.writeString(grave.dimensionName);
+                buf.writeInt(itemCount);
+                buf.writeInt(i - 1);
+                buf.writeByte(grave.availability);
+            }
+
+            responseSender.sendPacket(PacketIdentifiers.PLAYER_GRAVES_GUI, buf);
+        });
+        ServerPlayNetworking.registerGlobalReceiver(PacketIdentifiers.SINGLE_GRAVE_GUI, (server, player, handler, serverBuf, responseSender) -> {
+            NbtCompound nbtProfile = serverBuf.readNbt();
+
+            if (nbtProfile == null) return;
+
+            GameProfile profile = NbtHelper.toGameProfile(nbtProfile);
+            if (profile == null) return;
+
+            UUID graveId = serverBuf.readUuid();
+
+            YigdConfig config = YigdConfig.getConfig();
+            YigdConfig.GraveKeySettings keySettings = config.utilitySettings.graveKeySettings;
+
+            List<DeadPlayerData> data = DeathInfoManager.INSTANCE.data.get(profile.getId());
+            if (data == null) return;
+
+            DeadPlayerData graveData = null;
+            for (DeadPlayerData grave : data) {
+                if (!grave.id.equals(graveId)) continue;
+
+                graveData = grave;
+                break;
+            }
+
+            if (graveData == null) return;
+            NbtCompound graveNbt = graveData.toNbt();
+
+            PacketByteBuf buf = PacketByteBufs.create();
+            buf.writeNbt(graveNbt);
+            buf.writeBoolean(keySettings.enableKeys && keySettings.getFromGui);
+            buf.writeBoolean(config.graveSettings.unlockableGraves);
+            buf.writeBoolean(config.commandToggles.retrieveGrave && YigdCommand.hasPermission(player, "yigd.command.restore"));
+            buf.writeBoolean(YigdCommand.hasPermission(player, "yigd.command.delete"));
+            buf.writeBoolean(config.commandToggles.robGrave && YigdCommand.hasPermission(player, "yigd.command.rob"));
+
+            buf.writeBoolean(DeathInfoManager.INSTANCE.unlockedGraves.contains(graveId));
+            buf.writeBoolean(config.graveSettings.graveRobbing.tellRobber);
+
+            responseSender.sendPacket(PacketIdentifiers.SINGLE_GRAVE_GUI, buf);
         });
     }
 }