+++ src/main/java/com/b1n_ry/yigd/core/ServerPlayerEntityImpl.java
@@ -1,0 +1,7 @@
+package com.b1n_ry.yigd.core;
+
+import net.minecraft.util.math.Vec3d;
+
+public interface ServerPlayerEntityImpl {
+    Vec3d getLastGroundPos();
+}